Improved image extraction.
Never remove any query string argument on a resource that will be
downloaded. It does not matter we pass a _utm argument, and worse,
it shuffle the arguments on the URL (which should never matter but it
does with some CDN).

Transform <picture> elements as <img> with a bug srcset. It's not ideal
but it's an improvement.

After an image has been downloaded, check that's really an image we
can convert and serve in the browser (jpeg, png, gif, webp, ico, bmp
and tiff).