Added internal/email package and its configuration.
Email sending is controlled by Config.Email values.

If host is empty or debug is false, no mail can be sent. A function
configs.CanSendEmail() informs of this capability, and a template
variable "canSendEmail" reflects the result of this function.

When debug is true, messages are sent to stdout, for debuging and no
real email is sent.