Password recovery procedure.
If a user forgets their password, we provide a form where they can
enter their account's email address. Regardless of the mail being
present in the db, we send an email with further instructions.
If the email is indeed linked to an account, the message contains a
link (valid for 2 hours) where the user can set a new password.
When successful, the user can then go to their login page.

The link for recovery (and the HTTP routes) are not available if the
mail configuration is not set.

Note: we use a very naive in-memory k/v store to keep the recovery link
codes and expire them. While it works, we might have to move to
something distributed if one day there's a need to have multiple
readeck instances running behind a load balancer. The same applies for
the cancel processes that exist elsewhere.