New pkg/forms package
This will replace pkg/form.
Although the declarative way of doing forms works quite well it shows
some limits when we need to know if a field was set and/or is null.
The validation process of pkg/form was not optimal.

Instead, pkg/forms provides a form that must be explicitly created
during runtime. A form is nothing more than a list of field. And a
field is described by a simple interface providing a good latitude
of implementation.