Sends more headers when retrieving or creating a bookmark.
A new Link type lets you create (and send) a structured Link header
with a response.

When a bookmark is retrieve of created, we send a "Bookmark-Id" and
a "Link" (with rel=alternate, type=text/html) with the response.

Pagination links are now using the Link type to write their headers.