Receive a list of resources during bookmark creation.
Replaced the "html" entry in bookmark form with a "resource" list.
Each resource goes to a store providing a cache for the HTTP client
used by the extractor.

This let us send the full HTML page as we want it to be parsed and all
the remote resources that could be needed during extraction.