Added pkg/annotate
This package adds annotations to a DOM node based on a specific
selector format.
The containers are always Element nodes (not text nodes), with offsets
based on the full text content of said nodes.
This way, when we wrap annotation text nodes, the document's structure
is changed but any annotation is still valid.