Updated casbin matcher and policies.
The default role manager's matching function (KeyMatch2) was using
regular expressions. Which was not really needed and very slow the
moment you want to list all permissions for a given group.

By rewriting the subjects as path like in the policy, we can simply
use a path matching function as out "g" function.

Added two function in internal/acls:

- GetPermissions: returns the permissions for one or more groups
- InGroup: checks if a group contains all the permissions of another one