modernc/sqlite driver
This comes on top of mattn/go-sqlite3 driver. Only one of both can be
enabled at build time.
For performances reasons, it's better to use the CGO sqlite3 driver
but the native go version provides a very easy way for anyone to
compile and to provide cross compiled binaries.kkkkk