API documentation
This is the first commit with very few routes documentation.
The documentation source is divided in several files that are merged
together using file-compose, which provides a mechanism to merge parts
of yaml files together.

The documentation is served by RapiDoc. It's a true web component and
it provides some features for future improvements, should we need them.

Thank to a good styling system, and the use of variables, it was a few
lines of code to customize RapiDoc to Readeck's colors :)