Limit the title length on bookmark cards.
This is related to #18. We could shorten the text at the model's level
upon creation and update but, since titles can be changed, we can keep
storing them as we get them after extraction and only apply cosmetic
changes when we see them.

On a corpus of more than 2000 bookmarks from various sources, the
maximum length is 293 characters. With an median of 44.
A 90 character limit seems enough to cover a vast majority of cases.