Extract a bookmark's links later.
Fetching a page's link is a good idea and opens up the gate for cool
features but it clearly had an impact on the creation time.
This changes the whole process by fetching the links after the bookmark
has been saved.

- In the extract package, a new step "StepDone" is introduced. Unlike
  the others, it always runs, even when the process is canceled.
- Moved the database update and zip file creation to a processor running
  during this new step.
- Extract the links from the DOM in a earlier processor but don't
  fetch them yet, only save them in the extractor's context.
- One last processor performs the fetch after the bookmark was updated.

A next step could be to introduce a new bookmark state (like "partial")
and use it to force a refresh on the bookmark's page.