Changed application password authentication.
Instead of having to check every user's application password upon HTTP
authentication, we store the application passwords with a fixed salt
(per user). This provides the following workflow for application check:

- Retrieve the user from its username
- Create a password hash with the salt derived from the user ID
- Retrieve the application password (credential) for this user ID
  and hash.

There is still a time penalty using this authentication method but
it's a fixed one and it mitigates the risk of DDOS because of too many
application passwords for the same user.

Passwords are stored using the standard scheme so there is an upgrade
path when necessary.

closes #22