Improved the onboarding process
The terminal based onboarding was not very user friendly and quite
error prone, or not working at all in containers.

This introduces a web based onboarding with a single route to create
the initial user.

When there are no user in the database, a /onboarding route and any
authenticated route will redirect to it.

Once the first user is created, this route is not accessible anymore
and is not even mounted the next time Readeck starts.