Added pkg/zipfs ZipRW.
ZipRW is a simple wrapper around zip.Writer and zip.Reader.
It can be used as a simple zip writer to add files and call it a day.
Or, with an io.Reader source (being a zip content), it can copy files
from source to destination.
When used with io.ReadCloser or io.WriteCloser, source or destination
are closed when the whole ZipRW is closed.