Show article content for any bookmark type.
Before, we'd display the content only for bookmarks with an "article"
type. Since, we can soon inject HTML content from content scripts, we
want to read them.

This adds a "HasArticle" flag to a bookmarkItem and uses it almost
everywhere a test on "Type == article" was made.
When this flag is true, we always show the content, regardless of the
bookmark type.