Display the bookmark list of a collection during creation.
Both GET and POST on collection/add receive the api.withBookmarkList
middleware. We can then build a classic bookmark list with its
pagination.

The api.withBookmarkList can use GET and POST parameters when creating
the filter form. We need this to update the preview during the creation
of a collection, in case we stay on the same page (ie. the name triggers
an error)

In pkg/libjet, an attrList element can receive a boolean, resulting in
an value-less attribute.