Add a QR code to printed and epub bookmark.
On the print version of a bookmark, a QR code of the original article
is present at the top of the page.
On the print and epub version, a QR code of the link to the bookmark in
readeck is present at the end of the page (or chapter).

This moves a bit some css print declarations, in order to display the
bottom bookmark in the content's column.