Improved youtube transcript fetcher.
Before this change, the script would retrieve the first transcript
available, after a naive sort of non automatic transcripts first.
The API provides the information of the default transcript track and
we try to use this one when it's present. Otherwise, we rely on sorting
the transcript list and prioritize on non automatic, English first.