Bookmark import tooling.
This introduces the necessary interfaces, tasks and tools to import
bookmarks from various sources.

The import starts from an adapter that defines functions to load the
data and an iterator for the real data.

Each item from the iteration must provide a URL and can optionally
provide more data and even full content when it's available.

The background task will then proceed all the data and launch a
bookmark creation task for each item from the iteration.