Fixed a bug with uploaded data when creating a bookmark with multipart.
Creating a bookmark with multipart/form-data lets the client send
prefetched resources. These are stored in a cache during the extraction
process. The cache entry was implementing io.ReadCloser the wrong way,
resetting the content when reaching EOF. This can fly undetected until
you start using this reader with an io.TeeReader (to later recombine
with a multireader). If the content is small, it reaches the end, reset
and, if you recombine, you get a duplicated content.

This fixes the issue by extending bytes.Reader in cacheEntry and
implementing the io.ReadCloser by simply emptying its content.