Provide a site config overriding mechanism.
Introducing `$.overrideConfig()` for content scripts. This function
takes a site config (config) and a URL (src) and replaces the current
configuration with the one found for the given URL.

This let us load a configuration for sites acting as proxies for other
websites.

Added a content script doing exactly that for sites served by the
BNF newspaper "proxy".