Images in list items as inline-block
An image in a list item is often a replacement for a bullet. We can't
detect it  but at least it should not be displayed as a block to avoid
inelegant line breaks.

It seems to have no impact on illustration pictures in list item since
they're usually big enough to trigger a line break (or are followed by
a block such as a paragraph).