Try to keep images in figure blocks.
Readability uses regular expressions to filter elements by class and ids.
It works well enough but could remove images or block that contain classes
with "media" in their names.
Here, we can assume that if an image is in a figure tag, it convey an
editorial purpose and should be kept. It's the case, for example, with
mediapart.fr.