Improved video playback privacy and security.
So far, the playback of embed videos was made possible by a very lax
CSP rule `frame-src "*"`. This introduces a safer way to embed videos.
If we have an "Embed" value in the bookmark, we parse it and fetch the
iframe's src attribute. From there, we can retrieve the hostname that
becomes the only allowed host on the CSP rule.
Extra bonus: we can now force a Youtube iframe's src to the "nocookie"
variant.