Bookmark public link.
This lets a user share a bookmark, with its content, using a public link.

To generate the public link, we create a 16 byte packet with an expiry
date and the bookmark ID (both unsigned 64 bit integers).
This packet is then encrypted and converted to a base64 string, which
lets us create the URL "/@b/{value}".

On a request to "/@b/{value}", the value is decrypted and unpacked. If
the current time is before the expiry date, we retrieve the bookmark
(it must be fully loaded) and use it in a specific template.

On the frontend, a new "share" button opens a menu with 2 items:
- Download EPUB (replaces the previous button)
- Share by Link