Enable loading image links.
If we post a picture URL (jpeg, png, svg, etc.) we can intercept the
loaded content-type during the "StepBody", set the DocumentType to
"photo", a picture URL and a title. And we're done.

We moved the ExtractPicture processor from StepDom to StepFinish
so any picture set outside of the DOM step can be processed.

See #83