Set the Accept header when downloading images from the archiver.
When a request accepts text/html, some website, like Reddit, return an
HTML response when requesting an image.

This had been dealt with for previews and icons but not for images
in the articles. Now, while processing an img, picture or source
tag, we set an Accept header with "image/webp,image/svg+xml,image/*,*/*;q=0.8".

Not accepting text/html seems to be good enough but this way is more
"browser like".