Set colors with a palette generator.
The palette generator create a color scale with Tailwind's 11 levels
using an okhsl color space. It yields good and consistent results for
both hue and luminosity variations.

A scale is define as a css custom property,
ie. --color-gray: color-scale(#ccc) and generates properties like
--color-gray-N.
These custom properties are simple RGB values that you can use as,
for example, rgb(var(--color-gray-50)). This is how Tailwind set
a color (with a property for opacity).

The new scale is as close as possible as the previous one with minor
changes.

This introduces a dark scale that will be used after more changes to
templates and styles.