Send the CSP nonce with every Turbo request.
Before performing any Turbo request, we pass a "X-Turbo-Nonce" header.
The middleware settings the CSP intercepts and reuses it when it's
present. This let us execute the script that sets the "js", "no-js"
class on the body, even after a full Turbo page load, as it's the case
with bookmark loading.