Theme switcher button.
This adds a button in the main menu to switch between themes.
A script sets the theme (a class on the document) as soon as possible
to avoid a disgraceful FOUC effect.

The switcher button is linked to a controller that cycles between
auto / dark / light themes. The "auto" theme uses the OS preferences.

The theme is stored in the user's local storage and the switch is made
using only the "storage" event. This let us change the theme in every
opened tab or window.