Added an embed HLS video player.
A new /videoplayer?manifest= route can be used as an iframe source
to display a video player (like youtube or vimeo do). Something we can
and will use for captured content providing only a link to an
HLS manifest.

We'll use an iframe to avoid bloating the bookmark route with more JS
than needed, a specific CSP, etc.