Improve label sorting
On the label list, the order was determined by a simple ASCII sort,
which is problematic with accented characters and some others.

This introduces a collation that can be used directly or as a
collate statement with SQLite. It matches what postgresql already does
by default.

A migration takes care of reordering labels in bookmarks.