Simplified healthcheck command.
Performing an HTTP request is time consuming (not that much but still)
and can fail due to configuration constraints (server.allowed_hosts).
On top of that it fills the log on every healthcheck.

What we can do instead is only open a connection to the server's TCP
host and port. If the server is down, it will return an error.

See #119