Import from Wallabag v2
This one is a bit more involved and uses the Wallabag API.

The first step will validate the submitted API credentials and try
to authenticate.
Should this work, the background process will then retrieve bookmarks
by batch of 10 and process each one of them.

It **keeps** the original content as Wallabag saved it initially. It
might be that Readeck could perform better in some cases (or the other
way around) but as a user I expect my import to restore my data as much
as they were.