Added pkg/forms.FileField and pkg/forms.BindMultipart
forms.FileField implements a form field that can receive a "file" from
a multipart/form-data request. The field holds a FileOpener as a value
(defaults to a wrapper around multipart.FileHeader) and provides an
"Open" function to retrieve an io.ReadCloser from it.

forms.BindMultipart, binds every file entry in the multipart request
body to their respective FileField (only the last entry for the same
filename) and then proceeds to Unmarshal the values for regular fields.