Provide a custom error page when a bookmark does not exist.
We need that to fix #221. When a use removes a bookmark and stays on
the page until deletion, the turbo-frame content is replace with
"content missing".

This introduces a middleware that can set a custom template for a given
status code. The responseWriterInterceptor uses it when it renders
the error in HTML.

Changed the turbo-frame id to remove the bookmark's ID that wasn't
needed and isn't available in the error page's context.