Better description of youtube videos
We retrieve the full description from the API. Its first line becomes
the bookmark's description and, when there's more content in the full
description, we inject it as the beginning of the bookmark's HTML
content.

Then, we transform line breaks and URLs into paragraphs and links.

This forces readability in loose mode (only content cleanup) to keep
all the injected content.

Closes #261