UI: hide top navigation when scrolling down.
When a user starts scrolling down, a new controller adds a class to the
document's body (scrolled-down). The class is removed when scrolling up
again.

This lets us set a different position (hide) the top navigation when
scrolling and adapt other UI element when it's hidden.

This commit also adds smaller menu icons.