Increase the size of "Add A Label" placeholder
The "Add A Label" input box was previously annoted with `text-sm`. This set the
size of the placeholder text to .875rem, which is smaller than the default font
size of 16px. On mobile browsers, font sizes below 16px on input boxes cause the
browser to zoom in to center the input text box in the viewport. This is
intended to provide better visibility into input for web pages not designed for
mobile. The Readeck pages are optimized for mobile and look quite good on
mobile, it is quite jarring when the browser reshapes the viewport around this
little text box.

This change removes the `text-sm` class from the add label input box, so that
the text size reverts to the page default of 16 and mobile browsers do not zoom
in on the text box when you click on it. See the before/after pictures and
videos for more info.