Introducing trusted_proxies setting.
⚠️ breaking change but transition should be smooth

This removes the use_x_forwarded_* settings and the READECK_USE_X_FORWARDED
environment variable. It adds a new trusted_proxies setting that's a
list of IP range that are allowed to set X-Forwarded headers.

This intends to resolve the most common configuration issue.

trusted_proxies defaults to the following list:

- 127.0.0.0/8
- 10.0.0.0/8
- 172.16.0.0/12
- 192.168.0.0/16
- fd00::/8
- ::1/128

This provides a working out of the box deployment in 99% of cases.

The trusted proxy list can be changed in the configuration with
trusted_proxies (in server section) or READECK_TRUSTED_PROXIES
environment variable.