Save reading progress while scrolling the document.
This replaces the scroll-indicator controller with a scroll-progress
controller, placed at a level higher.

scroll-progress does the following:

- track scroll progress to live update an indicator (fast)
- track anchors during scroll and dispatch an event on changes (slower)

An intersection observer monitors elements as they enter a small window
located bellow the viewport's center. The first intersecting element
is then sent to a function that updates the current position and
anchor selector after a timeout.
On change, a second (longer) timeout updates the values and triggers
an event that can be picked up by the request controller.

The first timeout avoids picking up event too often while the second,
longer, one triggers the event only after a delay.

On page load, the initial position is restored if there is a tracked
element (that is, some text) and no URL fragment. It then does restore
a position when the user open a bookmark from the highlight list.

When an page doesn't contain a tracked element (no text), the bookmark
is considered viewed and the position is saved to 100%. This happens
on pictures and videos without transcript.

Unrelated:
- make the scroll-tracker scroll listener passive.
- push the URL to the history in scrollto controller.