Support setting secret key and email settings from environment variable.
Closes #348

Important note: starting Readeck for the first time
with `READECK_SECRET_KEY` leads to an empty configuration file.
This is normal and it works but, in this mode of operation, you must
provide the secret key from the environment each time you run Readeck.

You can still add parameters in the configuration file when needed,
including the secret key if you change your mind.