Fixed a bug with search string escaped backward slash.
Take a search query like "test\\"; the resulting string should be `test\`.
Instead it becomes `test\\`. This leads to wrong query if you have a
label that's named, say, `test\`.

This fixes the search query.

On the label lookup side, the label is now added to the search query
with proper quotes, using `strconv.Quote` (instead of a broken fmt.Sprintf).

This also fixes a bug with the label info API route that would incorrectly
returns a "href_bookmarks" value with a "label" query parameter, instead
of "labels".

Closes #349