Add an option to configure an initial admin user
This adds a new section to the config, `admin`, which lets one set up an
initial admin user right from the configuration, without having to visit
the onboarding page after launching Readeck.

The new options are also configurable through environment variables.

The password *must* be in passlib-format hashed form already, and will
be inserted into the table as-is.

Fixes #352.

Signed-off-by: Gergely Nagy <me@gergo.csillger.hu>