add word_count to bookmark api (#392)
the wordcount field is only exposed in the UI in various places. however, the
underlying field `word_count` is not exposed in the UI nor the API.

this PR exposes this field on the bookmarks api, to enable access for those wanting to get an the underlying metadata in their applications.

closes #392

Reviewed-on: https://codeberg.org/readeck/readeck/pulls/392
Reviewed-by: olivier <olivier@noreply.codeberg.org>
Co-authored-by: eleith <eleith@noreply.codeberg.org>
Co-committed-by: eleith <eleith@noreply.codeberg.org>