New bookmark export system.
Moved the bookmark export_* functions to internal/bookmarks/converter.
This module describes an interface that the export route can use to
export content in Markdown and EPUB formats.

The converter module provides the article retrieval from a bookmark.

The markdown exporter can export a single markdown file, with images
linked to the originating Readeck instance.
When the "Accept" header is multipart/alternative, Readeck sends a
multipart response with each bookmark converted to markdown and all
the images streamed along.

This new markdown exporter could be used by the incoming Obsidian plugin.
See #408