New bookmark filtering tools.
Bookmark list filters originated as a form that was later added into
the collection form, with a lot of redundancies here and there.

While the situation got better with new forms, this provides a clean
and common bookmarks. Filters struct that can be initialized from a form
and can update any given form.

The filter request is created by the Filters struct.