Fixed layout and update after highlighting.
reverts 55dc7bc

When trying to resolve #407, I removed "container-type: inline-size;"
on the .bookmark-content class. It was a massive mistake that broke
the layout with content that's overflowing (with tables).

Back to the drawing board, adding a method="morph" to the article
turbo stream does the trick and keeps a properly working layout.