New Key Material
This replaces the (bad) secret key interpolation with a set of keys
derived from the secret key, using HKDF.

- Use pkg/securecookie and pkg/csrf with configs.Keys
- Application password are hashed using a derived key with the user uid
  in its context.
- API Token are simplified (MAC signed using a specific derived key)
- Shared bookmark links are encrypted using blowfish + backed2b (EtM) in
  a 160-bit block

As this is a breaking change, users can now renew their application
passwords.

API Token can be retrieved on the given page. From the browser extension
a logout / login is necessary.