Refactored migration 16
Perform one query per update instead of a big case values. It's slower
but more sequential.

Rename bookmark files by creating hard links of the original ones in a
new folder and, when everything succeeded, remove the old folder and
rename the new one. If this fails, the whole transaction fails and
the original folder still intact.