Fixed (again) Safari's highlight issue
For good this time, without any fragile hack, only CSS.
The already known root cause of this bug was "container-type: inline-size;"
on bookmark-content. It would trigger a scroll top after refreshing the
article's contents (hence, after a highlight).
This "inline-size" was necessary though, to avoid issues with overflowing
content (ie. an article with a large table).
Turns out that "contain: inline-size" provides the needed feature
without the Safari bug on refresh.