Added OriginalLinkProcessor
This extract processor checks if a Link: <url>; rel="original" is present
in the extracted page.
When it finds one, it changes the extracted URL to its original value.
This headers is used, at least, by archive.md and archive.org.

Special case when the header is present and we have a "Readeck-Original"
header matching the found URL. Then the whole process is restarted anew
with said URL. This combination exists only for Readeck's publicly
shared bookmarks.