Improved Pocket import adapter.
The adapter assumed that the CSV file contained in Pocket's zip file
had a fixed number of columns. It seems to vary and/or has changed.

This uses the header labels to find the right columns and should last
as long as Pocket doesn't make more radical changes to their export
format.