Read the server port from the config file in the container unless explicitly overridden
If the container sets the environment variable, it becomes impossible to override the port via the configuration file because environment variables have higher precedence. Overriding the host makes sense, so that readeck is reachable from outside the container by default. However, fixing the port to the default serves is not required and can be confusing during deployment.