Handle annotation contiguity.
An annotation that start right after a given tag might actually start
inside (at the end of the previous sibling). For example:

<p>test <strong>foo<strong>start</p>
                   ^            ^

In the end, we end up with only "start" being selected but we could
(and should) start the range right before start, outside the "strong"
element.

This turns to be crucial when wrapping content with annotations for
we receive an overlap error when multiple annotations are contiguous
and one tries to start at the very end of the previous one.

This resolves #497.