Added a reminder about the browser extension.
Using the mechanism introduced in the extension v2.4.5 to handle the
message it sends to advertise it's been installed and configured,
we can determine that the extension is not installed.
https://codeberg.org/readeck/browser-extension/commit/4e2d0af4a2c0529ee76249a6dc37da2d7fcbd4e0

When the extension is not installed (or not configured for Readeck's
host), we show a reminder in the sidebar so a user can learn about the
extension.

Added the same text and link on the onboarding page.